// rapportini/src/app/api/models/document-type-dto.ts
/* tslint:disable */
export interface DocumentTypeDto {
  code?: null | string;
  deleted?: null | boolean;
  description?: null | string;
  extCode?: null | string;
  father?: null | DocumentTypeDto;
  groupType?: null | string;
  identity?: null | string;
  legalName?: null | string;
  sectional?: null | string;
}